Added a html functionality to add a new item

+++ resources/web.py
@@ -35,12 +35,33 @@ def item(store_id):
 @blp.route('/home/store/<store_id>/item/delete/<item_id>', methods=['GET', 'POST'])
 def item_delete(store_id, item_id):
     if request.method == 'POST':
-        item = ItemModel.query.get_or_404(item_id)
-        db.session.delete(item)
-        db.session.commit()
-        return redirect(url_for('Web.item', store_id=store_id))
-        #return render_template('index.html', user=url_for('Web.item', store_id=store_id))
+        if request.form['item_submit'] == 'Delete Item':
+            item = ItemModel.query.get_or_404(item_id)
+            db.session.delete(item)
+            db.session.commit()
+            return redirect(url_for('Web.item', store_id=store_id))
+        #return render_template('index.html', user=request.form['item_submit'])
         #return redirect(request.url)
+
+@blp.route('/home/store/<store_id>/item/add', methods=['GET', 'POST'])
+def item_add(store_id):
+    if request.method == 'POST':
+        if request.form['item_submit'] == 'Add Item':
+            item_data = {
+                "name": request.form.get('item_add_name'),
+                "price": request.form.get('item_add_price'),
+                "store_id": store_id,
+                "description": request.form.get('item_add_desc')
+            }
+            item = ItemModel(**item_data)
+            
+            try:
+                db.session.add(item)
+                db.session.commit()
+                return redirect(url_for('Web.item', store_id=store_id))
+            except SQLAlchemyError:
+                abort(500, message='An error occurred while inserting the item.')
+
 
 #def get_items(store_id):
 #    items = ItemModel.query.filter(ItemModel.store_id==store_id).order_by(ItemModel.price.asc())
